Fix broken initialization in CG
If `x0=None` is supplied as an argument to `pylops.optimization.solver.cg` (which seems to be intended) `x` is initialized as the zero vector. In this case the residual vector `r` should be `r = y - Op * x = y` and not `r = Op.T * y` (where * is matrix multiplication and .T is transposition).