Adds R-linear operators
Many useful linear operators are only linear over real numbers, and
complex numbers when they are treated as being elements of R^2
(e.g., (a+ib) -> (a, b)). These are known as R-linear operators, to
distinguish them from operators that are also linear over complex
numbers (C-linear operators). Examples include operators that
extract the real or imaginary components of the input, and
complex conjugation.

This commit introduces R-linear operators to PyLops. An operator
is specified to be only R-linear (and thus not C-linear) by
setting the new `clinear` attribute on the LinearOperator class
to be False.

When operators are combined, such as by composition or stacking,
the combined operator will be R-linear if any of the component
operators are. This commit implements this in the `dot` method
of the `LinearOperator` class, and also in the `VStack`, `HStack`,
and `BlockDiag` operators.

As complex values are considered to be elements of R^2 when
using R-linear operators, the dot product test for these operators
uses the R^2 inner product rather than the complex inner product.